+++ src/Circuit.hs
@@ -2,7 +2,8 @@ module Circuit
   ( module Circuit.Affine,
     module Circuit.Arithmetic,
     module Circuit.Lang,
-    module Circuit.Expr,
+    module Circuit.TExpr,
+    module Circuit.Compile,
     module Circuit.Dataflow,
     module Circuit.Solver,
   )
@@ -11,6 +12,7 @@ where
 import Circuit.Affine
 import Circuit.Arithmetic
 import Circuit.Dataflow
-import Circuit.Expr
+import Circuit.TExpr
+import Circuit.Compile
 import Circuit.Lang
 import Circuit.Solver
